refactor: rename & delete files
rename all files(lower case)

Combine check input modules into one file(check-user-inputs.js) & Delete them
