+++ src/constants/game-condition.js
