+++ src/controller/racing-car-game.js
@@ -1,8 +1,7 @@
 import RacingCarGameModel from '../model/racing-car-game-model.js';
 import RacingCarGameView from '../view/racing-car-game-view.js';
 import { $ } from '../modules/dom-selector.js';
-import checkUserCarNameInput from '../modules/checkUserCarNameInput.js';
-import checkUserRacingCountInput from '../modules/checkUserRacingCountInput.js';
+import { checkUserCarNameInput, checkUserRacingCountInput } from '../modules/check-user-inputs.js';
 import { SELECTOR } from '../constants/selector.js';
 
 export default class RacingCarGame {
