+++ src/modules/reset-input-text.js
@@ -1,4 +1,4 @@
-export default function initInputText(inputElement) {
+export default function resetInputText(inputElement) {
     inputElement.value = '';
     inputElement.focus();
 }
